requirements.txt
allure-pytest==2.8.7
allure-python-commons==2.8.7
attrs==19.3.0
importlib-metadata==1.4.0
more-itertools==8.1.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.4
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
zipp==2.1.0
